% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotly_network.R
\name{get_nodes}
\alias{get_nodes}
\title{get_nodes}
\usage{
get_nodes(obj, node.data = NULL, color = "color", size = "size",
  name = "names", type = "2d", merge = TRUE, ...)
}
\description{
get_nodes
}

